feat(main): define play screen and basic interactions (#6)
* refactor(mot): remove page for creating new Mots

New mots will be created on the same page that lists all the Mots. As
such, the 'new' page and route are no longer needed.

* style(mot): improve styling for Mot form

* refactor(mot): move Mot form to index page

* refactor(mot): define actions to create new Mots from index page

* feat(mot): remove form and item components for Mot model

Due to a design change (indicated in the README), the mot-form and
mot-item components are no longer needed. The form has been integrated
into the main page (mots/index.hbs). The application no longer uses Mot
items so the mot-item template is no longer necessary.

* feat(lastWord): define last-word component

The last-word component is used to maintain a reference to the last word
received from the server. This will be visible to all players.

* feat(index): integrate form and lastWord components in index page.

The form no longer has a separate component and has been integrated into
the index page. The lastWord component has also been integrated into the
page.

* feat(mot): install required libraries

- Install Bootstrap and Tether for UI elements
- Install Soundmanager for UI sounds
- Import sound files for UI sounds
- Import Roboto font

* feat(audio): create audio-player component

* chore(audio): initialize audio component

* feat(audio): import sound files

* docs(README): provide brief explanation of application