feat(menu): create menu button for executing actions
The 'action-button' is a menu button that allows the player to execute
an action within the game. An example of such an action would be
creating a new game.

There is only one action-button in use in the main menu, which is the
one used to create a new game.