refactor(menu): menu button can only link to pages
With the addition of the action-button, the menu-button now links to
pages and does nothing more. The 'Join Game' and 'How To Play' buttons
are menu buttons.

The change in function of the menu button required a change in the
styling. The menu button no longer changes colour in the mouseover
event. Instead, it changes colour based on the defined CSS for the
:hover property.