feat(game): join game room (#14)
* chore(style): import font-awesome icon library

* feat(user): create component to display players' names and scores

* feat(menu): define basic 'help' route

The 'help' route links to the help page where the user can view game
instructions.

* feat(menu): create page to allow user join a game session

The 'join' route links to the 'join' page where the user can input the
ID of a game to join it.

TODO: Prevent user from joining a room when it's full

* feat(menu): create menu button for executing actions

The 'action-button' is a menu button that allows the player to execute
an action within the game. An example of such an action would be
creating a new game.

There is only one action-button in use in the main menu, which is the
one used to create a new game.

* refactor(menu): menu button can only link to pages

With the addition of the action-button, the menu-button now links to
pages and does nothing more. The 'Join Game' and 'How To Play' buttons
are menu buttons.

The change in function of the menu button required a change in the
styling. The menu button no longer changes colour in the mouseover
event. Instead, it changes colour based on the defined CSS for the
:hover property.

* chore(style): define colours for player avatars and game contexts

* refactor(menu): change menu to reflect use of action and menu buttons

* feat(user): define component for displaying user info

The user-display component shows the user's avatar and score during a
game session.

* feat(variables): define service to store game information

* feat(websocket): add method to create a game on the socket server

* feat(menu): define 'join' and 'help' routes

* feat(game): initialize and store game variables when creating a new game

* feat(game): enable users join an existing game

- Manage game information by using the 'variables' service
- Store game information in each client's controller for easy access.
  This should be refactored into using the 'variables' service only
- Enable users join a game when the correct game ID is sent